--- a/log/log_test.go
+++ b/log/log_test.go
@@ -82,17 +82,15 @@ func TestWithConcurrent(t *testing.T) {
 }
 
 func BenchmarkDiscard(b *testing.B) {
-	logger := log.NewDiscardLogger()
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		logger.Log("k", "v")
+		log.Discard("k", "v")
 	}
 }
 
 func BenchmarkOneWith(b *testing.B) {
-	logger := log.NewDiscardLogger()
-	logger = log.With(logger, "k", "v")
+	logger := log.With(log.Discard, "k", "v")
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -101,7 +99,7 @@ func BenchmarkOneWith(b *testing.B) {
 }
 
 func BenchmarkTwoWith(b *testing.B) {
-	logger := log.NewDiscardLogger()
+	logger := log.Logger(log.Discard)
 	for i := 0; i < 2; i++ {
 		logger = log.With(logger, "k", "v")
 	}
@@ -113,7 +111,7 @@ func BenchmarkTwoWith(b *testing.B) {
 }
 
 func BenchmarkTenWith(b *testing.B) {
-	logger := log.NewDiscardLogger()
+	logger := log.Logger(log.Discard)
 	for i := 0; i < 10; i++ {
 		logger = log.With(logger, "k", "v")
 	}