--- a/log/value_test.go
+++ b/log/value_test.go
@@ -52,8 +52,7 @@ func TestValueBinding(t *testing.T) {
 }
 
 func BenchmarkValueBindingTimestamp(b *testing.B) {
-	logger := log.NewDiscardLogger()
-	logger = log.With(logger, "ts", log.DefaultTimestamp)
+	logger := log.With(log.Discard, "ts", log.DefaultTimestamp)
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -62,8 +61,7 @@ func BenchmarkValueBindingTimestamp(b *testing.B) {
 }
 
 func BenchmarkValueBindingCaller(b *testing.B) {
-	logger := log.NewDiscardLogger()
-	logger = log.With(logger, "caller", log.DefaultCaller)
+	logger := log.With(log.Discard, "caller", log.DefaultCaller)
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {