This change set does two simplifications:

1. Removes `NewDiscardLogger` in favour of a global `Discard` Logger as there is no need to have different instances of it.
2. Removes fast path from `BindValues` after observing a slight improvement in the benchmarks.