@ChrisHines 2015-04-26
I'm not a fan of the global Discard for three reasons.

1. It breaks naming symmetry with the other New*Logger functions.
2. It hurts discovery because it doesn't get grouped under the Logger type in the godoc index.
3. It becomes mutable and invite weird monkey patching ideas.

When I run the all benchmarks with the new BindValues I get:

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkJSONLoggerSimple           7155          7315          +2.24%
BenchmarkJSONLoggerContextual       10005         10080         +0.75%
BenchmarkDiscard                    380           381           +0.26%
BenchmarkOneWith                    737           914           +24.02%
BenchmarkTwoWith                    940           1275          +35.64%
BenchmarkTenWith                    2434          3574          +46.84%
BenchmarkPrefixLoggerSimple         2074          2365          +14.03%
BenchmarkPrefixLoggerContextual     3980          4526          +13.72%
BenchmarkValueBindingTimestamp      1181          1196          +1.27%
BenchmarkValueBindingCaller         2934          2684          -8.52%

benchmark                           old allocs     new allocs     delta
BenchmarkJSONLoggerSimple           19             20             +5.26%
BenchmarkJSONLoggerContextual       28             29             +3.57%
BenchmarkDiscard                    3              3              +0.00%
BenchmarkOneWith                    4              5              +25.00%
BenchmarkTwoWith                    4              5              +25.00%
BenchmarkTenWith                    4              5              +25.00%
BenchmarkPrefixLoggerSimple         5              6              +20.00%
BenchmarkPrefixLoggerContextual     11             12             +9.09%
BenchmarkValueBindingTimestamp      6              6              +0.00%
BenchmarkValueBindingCaller         6              6              +0.00%

benchmark                           old bytes     new bytes     delta
BenchmarkJSONLoggerSimple           812           842           +3.69%
BenchmarkJSONLoggerContextual       1159          1224          +5.61%
BenchmarkDiscard                    64            64            +0.00%
BenchmarkOneWith                    128           160           +25.00%
BenchmarkTwoWith                    192           256           +33.33%
BenchmarkTenWith                    704           1024          +45.45%
BenchmarkPrefixLoggerSimple         145           177           +22.07%
BenchmarkPrefixLoggerContextual     402           467           +16.17%
BenchmarkValueBindingTimestamp      192           192           +0.00%
BenchmarkValueBindingCaller         176           176           +0.00%
```

But it made me think about the problem and I think my take in #3 is better all around.