@tsenart 2015-04-26
Silly of me to have just ran `BenchmarkValue.*`. Obviously I missed the bigger picture :-)