@tsenart 2015-04-26
Regarding, symmetry of the `NewDiscardLogger`, I don't think it needs to be symmetric since it doesn't need more than one instance. Inspired from: http://golang.org/pkg/io/ioutil/#Discard