@ChrisHines 2015-04-26
I still prefer the way godoc groups the function version, but I'll defer to majority opinion. @peterbourgon, what do you think?