@peterbourgon 2015-04-26
I agree with @ChrisHines and prefer `func NewDiscardLogger() Logger` over `var Discard = LoggerFunc(...)`. I'd prefer clients/users to deal exclusively in first-class Logger objects with different implementations, than to have a separate package global.