@peterbourgon 2015-04-26
For a vague analog from the stdlib, see [ioutil.NopCloser](https://golang.org/pkg/io/ioutil/#NopCloser). For the record I'd prefer to stick with that naming scheme, too: `log.NewNopLogger`.