@tsenart 2015-04-26
OK, no problem. Now, we got talking about what should be public and what
shouldn't. If this Discard logger is only used in tests, should we really
export it?

On Sunday, April 26, 2015, Peter Bourgon <notifications@github.com> wrote:

> I agree with @ChrisHines <https://github.com/ChrisHines> and prefer func
> NewDiscardLogger() Logger over var Discard = LoggerFunc(...). I'd prefer
> clients/users to deal exclusively in first-class Logger objects with
> different implementations, than to have a separate package global.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/ChrisHines/gokit/pull/2#issuecomment-96424238>.
>


-- 
Sent from Gmail Mobile