@peterbourgon 2015-04-26
+1 to defining the discard logger only in `package log_test`