@tsenart 2015-04-26
Open source <3

On Sunday, April 26, 2015, Peter Bourgon <notifications@github.com> wrote:

> Huge thanks to you for your continued and thorough PRs! :)
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/ChrisHines/gokit/pull/2#issuecomment-96428338>.
>


-- 
Sent from Gmail Mobile